Create Question_Code_Finder.py
This program takes a list of question codes [strings] as a csv file and then looks for those question codes in all text files in the directory [text files of questionnaires]. It then compiles the resulting matrix into a new csv file with the first column consisting of question codes and subsequent columns showing "Si" or "No" depending on whether the question is in the text file. I wrote this program while working as a research assistant at the Latin American Public Opinion Project to generate a matrix showing which questions were asked in each country/year of the Americas Barometer, a poll of countries in the Americas.